// app-electronic-components-mobile/types.tsx
/**
 * Learn more about using TypeScript with React Navigation:
 * https://reactnavigation.org/docs/typescript/
 */

import { BottomTabScreenProps } from '@react-navigation/bottom-tabs';
import { CompositeScreenProps, NavigatorScreenParams } from '@react-navigation/native';
import { NativeStackScreenProps } from '@react-navigation/native-stack';

declare global {
  namespace ReactNavigation {
    interface RootParamList extends RootStackParamList {}
  }
}

export type RootStackParamList = {
  Auth: NavigatorScreenParams<AuthStackParamList> | undefined;
  Home: NavigatorScreenParams<HomeStackParamList> | undefined;
  Modal: undefined;
  NotFound: undefined;
};

export type RootStackScreenProps<Screen extends keyof RootStackParamList> = NativeStackScreenProps<
  RootStackParamList,
  Screen
>;

export type RootTabParamList = {
  Landing: NavigatorScreenParams<LandingStackParamList> | undefined;
  CategoryStack: NavigatorScreenParams<CategoryStackParamList> | undefined;
  Account: NavigatorScreenParams<AccountStackParamList> | undefined;
  CartStack: NavigatorScreenParams<CartStackParamList> | undefined;
};

export type RootTabScreenProps<Screen extends keyof RootTabParamList> = CompositeScreenProps<
  BottomTabScreenProps<RootTabParamList, Screen>,
  NativeStackScreenProps<RootStackParamList>
>;

export type AuthStackParamList = {
  Login: undefined;
  Register: undefined;
  Register2: undefined;
};

export type HomeStackParamList = {
  Root: NavigatorScreenParams<RootTabParamList> | undefined;
  Modal: undefined;
  NotFound: undefined;
  ProductDetail: undefined;
}

export type LandingStackParamList = {
  Landing: undefined;
}

export type CategoryStackParamList = {
  Category: undefined;
  All: undefined;
}

export type AccountStackParamList = {
  Account: undefined;
}

export type CartStackParamList = {
  Cart: undefined;
  Payment: undefined;
  EditAddress: undefined;
  PaymentSuccess: undefined;
}